Got the mesh rendering correctly
Fixed a bug where the indices of the mesh were in the wrong order so its winding order normal was pointing downwards.

Also fixed a bug where the end X triangles index was getting wrapped back to the start of the mesh X position, causing it to look like there were actually two meshes because it stretched all the way back to the start from one end.

+++ OpenGL Graphics Programming/Camera.cpp
@@ -145,8 +145,8 @@ void Camera::ProcessInput()
 			SetPosition(glm::vec3(tempVec));
 		}
 
-		inputManager.g_mousePosDifX *= m_mouseSens;
-		inputManager.g_mousePosDifY *= m_mouseSens;
+		inputManager.g_mousePosDifX *= (int)m_mouseSens;
+		inputManager.g_mousePosDifY *= (int)m_mouseSens;
 
 		m_camYaw += inputManager.g_mousePosDifX;
